GOV-7000 - asset get api using int data type instead of bigint resulting in errors on some numbers that display correctly in govern.

+++ d360.web/Controllers/V2/AssetsController.cs
@@ -105,7 +105,7 @@ namespace d360.web.Controllers.V2
                 case "DateTime":
                     return "datetime";
                 case "Number":
-                    return "int";
+                    return "bigint";
                 case "Decimal":
                     return "float";
                 case "Boolean":
